Extract shared items for formatters
