+++ lib/formatters/eslint.js
@@ -1,28 +1,11 @@
-const { red, green, yellow, bold, underline, dim, reset } = require('chalk');
+const { underline, dim, reset } = require('chalk');
 const stripAnsi = require('strip-ansi');
 const table = require('text-table');
 const { cwd } = require('process');
 const { relative } = require('path');
+const { status, summary } = require('./shared');
 
 const sourceRoot = cwd();
-
-const status = Object.freeze({
-  error: {
-    color: red,
-    icon: '\u2716',
-    severity: 'error'
-  },
-  success: {
-    color: green,
-    icon: '\u2714',
-    severity: 'success'
-  },
-  warning: {
-    color: yellow,
-    icon: '\u26A0',
-    severity: 'warning'
-  }
-});
 
 const tableOptions = Object.freeze({
   align: ['r', 'l'],
@@ -30,13 +13,6 @@ const tableOptions = Object.freeze({
     return stripAnsi(str).length;
   }
 });
-
-const summary = (problemTotal, sourceTotal) =>
-  bold.underline(
-    `ESLint found ${problemTotal} ${
-      problemTotal === 1 ? 'problem' : 'problems'
-    } in ${sourceTotal} source ${sourceTotal === 1 ? 'file' : 'files'}\n`
-  );
 
 const formatSource = (path, color, icon) =>
   `${color(icon)}  ${underline(path)}\n`;
@@ -76,6 +52,10 @@ module.exports = function (results) {
   });
 
   return reset(
-    `${summary(problemCount, results.length)}\n${successes}\n${problems}`
+    `${summary(
+      'ESLint',
+      problemCount,
+      results.length
+    )}\n${successes}\n${problems}`
   );
 };
