+++ lib/formatters/stylelint.js
@@ -1,15 +1,9 @@
-const { bold, reset } = require('chalk');
-
-const summary = (problemTotal, sourceTotal) =>
-  bold.underline(
-    `Stylelint found ${problemTotal} ${
-      problemTotal === 1 ? 'problem' : 'problems'
-    } in ${sourceTotal} source ${sourceTotal === 1 ? 'file' : 'files'}\n`
-  );
+const { reset } = require('chalk');
+const { summary } = require('./shared');
 
 /**
  * @type {import('stylelint').Formatter}
  */
 module.exports = function (results /*, returnValue*/) {
-  return reset(`${summary(0, results.length)}\n`);
+  return reset(`${summary('Stylelint', 0, results.length)}\n`);
 };
